requirements.txt
fastapi[standard]
SQLAlchemy
psycopg2-binary
pyjwt
requests
bcrypt==4.0.1
passlib==1.7.4

pytest
pytest-cov